% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Cumsum}
\alias{Cumsum}
\title{Cumulative sum, dropping NAs}
\usage{
Cumsum(x)
}
\arguments{
\item{x}{a numeric vector containing the values to be cumulatively summed.}
}
\value{
A vector of \code{length(x)}.
}
\description{
An Rcpp implementation of \code{\link[base]{cumsum}} that quietly discards
\code{NA}s. If an NA is encountered, the previous value is assumed. Also note
that the first value in \code{x} should not be \code{NA}.
}

