% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_sampling.R
\name{deltat_guess}
\alias{deltat_guess}
\title{Guess data sampling frequency}
\usage{
deltat_guess(time.x)
}
\arguments{
\item{time.x}{numeric vector; timer values, in any units.}
}
\value{
A numeric (scalar) giving the typical difference between time points.
}
\description{
A convenience function for guessing the sampling frequency of a dataset.
}

