% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Diff}
\alias{Diff}
\title{Faster base::diff}
\usage{
Diff(x)
}
\arguments{
\item{x}{a numeric vector containing the values to be differenced.}
}
\value{
A vector of \code{length(x)}, the first value being 0. This behaviour
  differs from \code{\link[base]{diff}}, where the length would be
  \code{length(x) - 1} (omitting the zero).
}
\description{
A more efficient implementation of base::diff(x, lag = 1, differences = 1).
}

