% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_correct.R
\name{elevation_correct}
\alias{elevation_correct}
\title{Elevation correction}
\usage{
elevation_correct(lon, lat, country)
}
\arguments{
\item{lon, lat}{numeric vectors (of the same length) giving positional
coordinates in units of (decimal) degrees.}

\item{country}{character; the three letter ISO code for the country in which
the \code{lon,lat} data were collected. See \code{raster::getData("ISO3")}.}
}
\value{
A numeric vector of the same length as \code{lon,lat} of elevation
  values (i.e. metres above sea level).
}
\description{
Extracts (reliable) elevation data from raster "alt" dataset,
  given positional coordinates (\code{lon,lat}). Requires the \code{raster}
  package to be installed. Essentially a wrapper for
  \code{\link[raster]{getData}} and \code{\link[raster]{extract}}.
}
\details{
Note that raster data may need to be downloaded, and hence an
  internet connection will be required if these files do not already exist in
  the working directory.

  Also note that elevation data are interpolated from the values of the four
  nearest raster cells. See the \code{method} argument in
  \code{\link[raster]{extract}}.
}
\examples{
\dontrun{
data(chaingang)

# Show the original elevation profile...
plot(elevation.m ~ distance.km, type = "l", data = chaingang)

chaingang <- transform(chaingang,
                       elev.new = elevation_correct(lon, lat, "GBR"))

lines(elev.new ~ distance.km, col = "red", data = chaingang)
}
}

