% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhmmss.R
\name{hhmmss}
\alias{hhmmss}
\alias{hms_to_sec}
\alias{sec_to_hms}
\title{HH:MM:SS conversion}
\usage{
hms_to_sec(hms)

sec_to_hms(sec, digits = 1, strip_zeros = FALSE)
}
\arguments{
\item{hms}{character string; a time in the form "HH:MM:SS" ("HH:" is
optional).}

\item{sec}{numeric; seconds values.}

\item{digits}{numeric (scalar); number of decimal places to be used for the
seconds value in the "hours:minutes:seconds" string.}

\item{strip_zeros}{logical; remove leading zeros? E.g. "00:26:01" becomes
"26:01".}
}
\description{
Convert between seconds and the more readable "HH:MM:SS" time format.
}
\examples{
x <- c("00:21:05", "25:51", NA, "00:26:01.1", "01:05:02.0")
(x <- hms_to_sec(x))
(x <- sec_to_hms(x, strip_zeros = TRUE))
}

