% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs, rhs}{an R object and function to apply to it.}
}
\description{
See \code{\link[magrittr]{pipe}} and/or \code{\link[dplyr]{chain}}.
}

