% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_sampling.R
\name{uniform_sampling}
\alias{uniform_sampling}
\title{Make rows uniform}
\usage{
uniform_sampling(.data, time_column, deltat = NULL, empty_fill = 0L)
}
\arguments{
\item{.data}{data.frame; the data to be made uniform on the basis of
\code{time_column}.}

\item{time_column}{identifier for the column in \code{data} that gives
(cumulative) time values. Right-hand-sided formulas will be deparsed,
otherwise this argument is passed to \code{'[['} as-is.}

\item{deltat}{numeric (scalar); the \emph{typical} sampling frequency of the
data. If \code{NULL} (default), a guess is made via
\code{\link{deltat_guess}}.}

\item{empty_fill}{the value with which to "fill" new rows in the returned
data. For most purposes this should be either \code{NA} or a numeric value.}
}
\description{
A useful function for working with data that aren't sampled at
  regular intervals and(or) contain gaps. This can present issues when, for
  example, trying to apply time-windowed rolling operations.
}
\details{
For time-windowed rolling operations, empty_fill should logically be
  set to \code{0L}; the use of NAs could generate large runs of NAs, causing
  complications.
}
\examples{
gappy_data <- data.frame(
  timer.s = c(1:100, 201:300),
  x = rnorm(200, 500, 30),
  y = rnorm(200, 1000, 300)
)
unif <- uniform_sampling(gappy_data, ~timer.s, empty_fill = NA)
head(unif)
}

